requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-openai
faiss-cpu
pypdf
python-dotenv
